refactor: address review comments on mock-LLM model-switch test
- Import APIRequestContext from top-level @playwright/test import
  instead of inline import() type references (review #1)
- Extract setChatInput to shared mock-llm-helpers.ts and reuse it
  in both mock-llm-conversation and mock-llm-model-switch (review #2)
- Add upstream reference for the padding trajectory entry coupling
  to CondensationMixin (review #3)
- Use direct field assertion on switchProfileBody.profile_name
  instead of loose JSON.stringify substring match (review #4)
- Remove redundant toBeVisible check after waitForNonUserMessageText
  already polls model-messages elements (review #5)
