+++ tests/e2e/mock-llm/utils/mock-llm-helpers.ts
@@ -380,5 +380,31 @@ export async function resetMockLLM(request: APIRequestContext) {
   expect(resp.ok(), `Reset mock LLM: ${resp.status()}`).toBe(true);
 }
 
+/**
+ * Set contentEditable chat input text and dispatch an input event.
+ *
+ * contentEditable divs don't respond reliably to Playwright's .fill() or
+ * .type(), so we set the text programmatically via page.evaluate().
+ */
+export async function setChatInput(page: Page, text: string) {
+  await page.evaluate(
+    ({ testId, inputText }) => {
+      const el = document.querySelector(`[data-testid="${testId}"]`);
+      if (!(el instanceof HTMLElement))
+        throw new Error("Chat input not found");
+      el.focus();
+      el.textContent = inputText;
+      el.dispatchEvent(
+        new InputEvent("input", {
+          bubbles: true,
+          data: inputText,
+          inputType: "insertText",
+        }),
+      );
+    },
+    { testId: "chat-input", inputText: text },
+  );
+}
+
 // Mock automation helpers removed — the automation test now hits the real
 // automation backend running inside the bin/agent-canvas.mjs stack.
